Add non-ESP32 Arduino core portability
- Centralize environment/time-source detection in platform_detect.hpp
- Add a non-ESP32 Arduino core time fallback that avoids <thread>/<chrono>
- Use the 32-bit timestamp format on the Arduino time fallback
- Fall back to a no-op TL_ASSERT where the target lacks a usable assert
  hook (expected.hpp)
- Quiet M5Utility warning output
- Declare library.properties architectures=* instead of esp32-only
- Guard IRAM_ATTR against redefinition when both the Arduino-core and
  espidf paths already provide it

+++ src/m5_utility/crc.hpp
@@ -90,7 +90,7 @@ public:
             uint_fast8_t cnt{8};
             while (cnt--) {
                 if (crc & 0x80) {
-                    crc = (crc << 1) ^ polynomial;
+                    crc = static_cast<uint8_t>((static_cast<uint16_t>(crc) << 1) ^ polynomial);
                 } else {
                     crc <<= 1;
                 }
@@ -184,11 +184,11 @@ public:
         while (len--) {
             uint8_t e{refIn ? reverseBitOrder(*data) : *data};
             ++data;
-            crc ^= (e << 8);
+            crc ^= static_cast<uint16_t>(static_cast<uint16_t>(e) << 8);
             uint_fast8_t cnt{8};
             while (cnt--) {
                 if (crc & 0x8000) {
-                    crc = (crc << 1) ^ polynomial;
+                    crc = static_cast<uint16_t>((static_cast<uint32_t>(crc) << 1) ^ polynomial);
                 } else {
                     crc <<= 1;
                 }
